+++ packages/orq-rc/src/orq_ai_sdk/models/getbudgetop.py
@@ -154,5 +154,5 @@ class GetBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-20T13:32:34.206Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-21T02:09:20.541Z")
     r"""The date and time the resource was last updated"""