+++ packages/orq-rc/RELEASES.md
@@ -3439,3 +3439,13 @@ Based on:
 - [python v3.14.0-rc.39] packages/orq-rc
 ### Releases
 - [PyPI v3.14.0-rc.39] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.39 - packages/orq-rc
+
+## 2025-10-21 02:11:08
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.637.3 (2.727.9) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.14.0-rc.40] packages/orq-rc
+### Releases
+- [PyPI v3.14.0-rc.40] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.40 - packages/orq-rc